+++ src/index.js
@@ -2,6 +2,7 @@ import './style.css';
 import displayUI from './modules/displayUI.js';
 import createPopup from './modules/popup.js';
 import { postLikes } from './modules/postLikes.js';
+import commenting from './modules/formComment.js';
 
 displayUI();
 
@@ -14,5 +15,8 @@ document.addEventListener('click', async (e) => {
     if (done.status === 201) {
       displayUI();
     }
+  }
+  if (e.target.classList.contains('formSubmit')) {
+    commenting(e.target.id);
   }
 });
